# collection-ui/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/users/{id}/board-games": {
            "post": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "addToCollection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CollectionEntryRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserBoardGame"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserCollectionResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/users/{id}/board-games/{gameId}": {
            "delete": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "removeFromCollection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "gameId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "BoardGame": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "yearPublished": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "minPlayers": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "maxPlayers": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "playTimeMinutes": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "weight": {
                        "type": "number"
                    },
                    "designers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BoardGameDesigner"
                        },
                        "uniqueItems": true
                    },
                    "publishers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BoardGamePublisher"
                        },
                        "uniqueItems": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "BoardGameDesigner": {
                "type": "object",
                "properties": {
                    "designerId": {
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
            "BoardGamePublisher": {
                "type": "object",
                "properties": {
                    "publisherId": {
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
            "CollectionEntryRequest": {
                "type": "object",
                "properties": {
                    "boardGameId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "condition": {
                        "type": "string"
                    }
                }
            },
            "UserBoardGame": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "userId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "boardGameId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "condition": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "email": {
                        "type": "string"
                    },
                    "passwordHash": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "UserCollectionResponse": {
                "type": "object",
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "collection": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BoardGame"
                        }
                    }
                }
            }
        }
    }
}